Update farsiteAPI.cpp
Updated lines for precipitation but just for metric system line as wrf output is in metric system (i.e. in millimeters,mm). The precipitation in output is total accumulated precipitation from the start of the wrf run. So to get hourly precipitation we need to subtract the accumulated precipitation at the current time step (or current file) with the value at previous time step (or the previous hour file). To see the changed lines just  cntrl-F and search for "Amit".